[Synthetics] Overview cards fix space change loading !! (#231177)
## Summary

fixes https://github.com/elastic/kibana/issues/224923

On changing spaces, list loads properly now

<img width="1728" height="891" alt="image"
src="https://github.com/user-attachments/assets/c5ba8b39-41ec-46c4-b7c3-a8f9fe993899"
/>

---------

(cherry picked from commit 1422b592cc3b67f6b0a1066256572ea0ac5a8b90)

# Conflicts:
#	x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/components/common/components/embeddable_panel_wrapper.tsx
#	x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/components/monitors_page/overview/overview/overview_grid.tsx
