+++ x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/components/common/components/add_to_dashboard.tsx
@@ -96,10 +96,12 @@ export const useAddToDashboard = ({
 
 export const AddToDashboard = ({
   type,
+  isLoading,
   asButton = false,
 }: {
   type: typeof SYNTHETICS_STATS_OVERVIEW_EMBEDDABLE | typeof SYNTHETICS_MONITORS_EMBEDDABLE;
   asButton?: boolean;
+  isLoading?: boolean;
 }) => {
   const { view } = useSelector(selectOverviewState);
 
@@ -127,6 +129,7 @@ export const AddToDashboard = ({
           data-test-subj="syntheticsEmbeddablePanelWrapperButton"
           iconType="dashboardApp"
           onClick={() => setDashboardAttachmentReady(true)}
+          isLoading={isLoading}
         >
           {i18n.translate('xpack.synthetics.embeddablePanelWrapper.shareButtonLabel', {
             defaultMessage: 'Add to dashboard',
@@ -146,6 +149,7 @@ export const AddToDashboard = ({
                   defaultMessage: 'Add to dashboard',
                 }
               )}
+              isLoading={isLoading}
             />
           }
           isOpen={isPopoverOpen}
