+++ x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/components/common/components/embeddable_panel_wrapper.tsx
@@ -5,15 +5,9 @@
  * 2.0.
  */
 
-import React, { FC } from 'react';
-import {
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiPanel,
-  EuiProgress,
-  EuiSpacer,
-  EuiTitle,
-} from '@elastic/eui';
+import type { FC } from 'react';
+import React from 'react';
+import { EuiFlexGroup, EuiFlexItem, EuiPanel, EuiSpacer, EuiTitle } from '@elastic/eui';
 import { AddToDashboard } from './add_to_dashboard';
 import { SYNTHETICS_STATS_OVERVIEW_EMBEDDABLE } from '../../../../embeddables/constants';
 
@@ -49,7 +43,7 @@ export const EmbeddablePanelWrapper: FC<React.PropsWithChildren<Props>> = ({
               </EuiFlexItem>
               {isSyntheticsApp && (
                 <EuiFlexItem grow={false}>
-                  <AddToDashboard type={SYNTHETICS_STATS_OVERVIEW_EMBEDDABLE} />
+                  <AddToDashboard type={SYNTHETICS_STATS_OVERVIEW_EMBEDDABLE} isLoading={loading} />
                 </EuiFlexItem>
               )}
               {titleAppend && <EuiFlexItem grow={false}>{titleAppend}</EuiFlexItem>}
@@ -57,7 +51,6 @@ export const EmbeddablePanelWrapper: FC<React.PropsWithChildren<Props>> = ({
             <EuiSpacer size="s" />
           </>
         )}
-        {loading && <EuiProgress size="xs" color="accent" />}
         {children}
       </EuiPanel>
     </>
