+++ x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/state/overview_status/actions.ts
@@ -21,5 +21,4 @@ export const quietFetchOverviewStatusAction = createAsyncAction<
 >('quietFetchOverviewStatusAction');
 
 export const clearOverviewStatusErrorAction = createAction<void>('clearOverviewStatusErrorAction');
-export const clearOverviewStatusState = createAction<void>('clearOverviewStatusState');
 export const initialLoadReported = createAction<void>('initialLoadReported');
