+++ x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/state/overview_status/index.ts
@@ -11,7 +11,6 @@ import { OverviewStatusMetaData, OverviewStatusState } from '../../../../../comm
 import { IHttpSerializedFetchError } from '..';
 import {
   clearOverviewStatusErrorAction,
-  clearOverviewStatusState,
   fetchOverviewStatusAction,
   quietFetchOverviewStatusAction,
   initialLoadReported,
@@ -59,12 +58,6 @@ export const overviewStatusReducer = createReducer(initialState, (builder) => {
     .addCase(fetchOverviewStatusAction.fail, (state, action) => {
       state.error = action.payload;
       state.loading = false;
-    })
-    .addCase(clearOverviewStatusState, (state, action) => {
-      state.status = null;
-      state.loading = false;
-      state.loaded = false;
-      state.error = null;
     })
     .addCase(clearOverviewStatusErrorAction, (state) => {
       state.error = null;
